chore: remove invalid workflow_dispatch property
workflow_dispatch does not support the `branches` property.  Per the [docs](https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/events-that-trigger-workflows#workflow_dispatch), the workflow must exist in the default branch but since it is run manually, you chose which branch you want to run the workflow on when manually invoking the workflow.